Add `sourceMapURL` option to node tool
Currently, the cli allows a client to
override the assignment of the `sourceMappingURL`
comment from the value passed via `source-map` to the
value passed via `source-map-url`. This is necessary
because the `file` property of a sourcemap, which
according to the spec should specify the minified
source filename, should not be the same as the URL
at which to find the sourcemap for this file.

In practice, the `file` property is not often used,
so the two can often be assigned to the same thing,
but there is a need in some cases to assign these
different.

This change updates the node tool api to reflect
the cli in this respect, introducing an option
`sourceMapURL` that will override the assignment
of `outSourceMap` to the value in the `sourceMappingURL`
comment.