fix: invalid filenames in `sources` of source maps
When the parsed files have no associated filenames (eg: a string or array of strings was passed to the `minify` function), the `sources` array of the generated sourcemap must only contain null values.

Currently, it contains the stringified indexes of the array. For example, passing a string to `minify` results in a `sources` array of `["0"]`, which breaks the sourcemap.

Also, this PR adds the `sourceMap.json` option. When explicitly set to `false`, the sourcemap is *not* stringified.